+++ changelog.markdown
@@ -5,6 +5,7 @@
 # 4.3.3
 - fixed #131 drags mixed with scrolling (caused by chrome update)
 - fixed #129 drag stucks on no-container release
+- fixed #128 canBeAccepted dod not work in combination with isContainer
 - added options object as second argument to onInit (requested in #65)
 
 # 4.3.2 War of the Worlds